src/mqtt_consumer/requirements.txt
boto3==1.36.0
botocore==1.36.0
jmespath==1.0.1
paho-mqtt==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.0
six==1.17.0
urllib3==2.3.0
